third_party/blink/web_tests/platform/generic/fast/forms/select/listbox-focusable-items-expected.txt

option/optgroup in menulist can be made focusable by setting tabindex as long as they aren't hidden

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS isFocusable("optionA") is true
PASS isFocusable("optionB") is false
PASS isFocusable("optionC") is false
PASS isFocusable("optgroupD") is false
PASS isFocusable("optionE") is false
PASS isFocusable("optgroupF") is false
PASS isFocusable("optionG") is false
PASS isFocusable("optionH") is true
PASS successfullyParsed is true

TEST COMPLETE

A
B
C
E
G
H
